% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{check_tokens}
\alias{check_tokens}
\title{Check student tokens}
\usage{
check_tokens(userfile, server = Sys.getenv("git.server",
  "https://gitlab.com/api/v4/"))
}
\arguments{
\item{userfile}{A csv file with student information (i.e., username and token)}

\item{server}{The gitlab API server}
}
\description{
Check student tokens
}
\details{
See \url{https://github.com/vnijs/gitgadget} for additional documentation
}
