% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{checkerr}
\alias{checkerr}
\title{Check error status}
\usage{
checkerr(code)
}
\arguments{
\item{code}{Code returned by an API request}
}
\description{
Check error status
}
\details{
See \url{https://github.com/vnijs/gitgadget} for additional documentation
}
