% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{collect_work}
\alias{collect_work}
\title{Create merge requests for each student/team}
\usage{
collect_work(
  token,
  assignment,
  userfile,
  type = "individual",
  server = "https://api.github.com/"
)
}
\arguments{
\item{token}{GitHub token}

\item{assignment}{Name of the assignment (e.g., "class345/class345-assignment1")}

\item{userfile}{A csv file with student information (i.e., username, token, and email)}

\item{type}{Individual or Team work}

\item{server}{The GitLab API server}
}
\description{
Create merge requests for each student/team
}
\details{
See \url{https://github.com/vnijs/gitgadget} for additional documentation
}
