% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{fetch_work}
\alias{fetch_work}
\title{Fetch all merge requests as local branches and link to a remote}
\usage{
fetch_work(token, groupname, assignment, pre = "",
  server = "https://gitlab.com/api/v4/")
}
\arguments{
\item{token}{Gitlab token}

\item{groupname}{Group containing the assignment}

\item{assignment}{Name of the assigment}

\item{pre}{Pre-amble for the assignment name, usually groupname + "-"}

\item{server}{The gitlab API server}
}
\description{
Fetch all merge requests as local branches and link to a remote
}
\details{
See \url{https://github.com/vnijs/gitgadget} for additional documentation
}
