% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadget.R
\name{gitgadget_callr}
\alias{gitgadget_callr}
\title{Launch gitgadget in a separate process}
\usage{
gitgadget_callr()
}
\description{
Launch gitgadget in a separate process
}
\details{
Using the \code{callr} package to launch gitgadget in a separate process so
  the console is not blocked. Rstudio viewer is used if available. See
  \url{https://github.com/vnijs/gitgadget} for documentation

 @importFrom callr r_bg
}
