% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadget.R
\name{gitgadget_url}
\alias{gitgadget_url}
\title{Start gitgadget and show url to open the application in an external browser}
\usage{
gitgadget_url()
}
\description{
Start gitgadget and show url to open the application in an external browser
}
\details{
See \url{https://github.com/vnijs/gitgadget} for documentation
}
