% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{read_ufile}
\alias{read_ufile}
\title{Reach user file}
\usage{
read_ufile(userfile, cols = c("userid", "team", "token"))
}
\arguments{
\item{userfile}{File with student information}

\item{cols}{Column names that must exists in the file}
}
\description{
Reach user file
}
\details{
See \url{https://github.com/vnijs/gitgadget} for additional documentation
}
