% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{remove_project}
\alias{remove_project}
\title{Remove a project}
\usage{
remove_project(token, id, server)
}
\arguments{
\item{token}{GitLab token}

\item{id}{Project ID}

\item{server}{The gitlab API server}
}
\description{
Remove a project
}
\details{
See \url{https://github.com/vnijs/gitgadget} for additional documentation
}
