% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{remove_users_repo}
\alias{remove_users_repo}
\title{Remove users from a repo}
\usage{
remove_users_repo(token, repo, userfile, server = "https://api.github.com/")
}
\arguments{
\item{token}{GitLab token}

\item{repo}{Repo the update}

\item{userfile}{A csv file with student information (i.e., username, token, and email)}

\item{server}{The GitLab API server}
}
\description{
Remove users from a repo
}
\details{
See \url{https://github.com/vnijs/gitgadget} for additional documentation
}
