% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{assign_work}
\alias{assign_work}
\title{Assign work to each student/team by creating a fork of the main repo}
\usage{
assign_work(
  token,
  groupname,
  assignment,
  userfile,
  tafile = "",
  type = "individual",
  pre = "",
  server = "https://api.github.com/"
)
}
\arguments{
\item{token}{GitLab token}

\item{groupname}{Group to create on GitLab (defaults to user's namespace)}

\item{assignment}{Name of the assignment to assign}

\item{userfile}{A csv file with student information (i.e., username, token, and email)}

\item{tafile}{A optional csv file with TA information (i.e., username, token, and email)}

\item{type}{Individual or Team work}

\item{pre}{Pre-amble for the assignment name, usually groupname + "-"}

\item{server}{The GitLab API server}
}
\description{
Assign work to each student/team by creating a fork of the main repo
}
\details{
See \url{https://github.com/vnijs/gitgadget} for additional documentation
}
