% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{collect_work}
\alias{collect_work}
\title{Create merge requests for each student/team}
\usage{
collect_work(username, password, groupname, assignment, userfile,
  type = "individual", pre = "", server = "https://gitlab.com/api/v3/")
}
\arguments{
\item{username}{Gitlab username}

\item{password}{Gitlab password}

\item{groupname}{Group containing the assignment}

\item{assignment}{Name of the assigment. file.path(directory, assignment) should exist}

\item{userfile}{A csv file with student information (i.e., username and token)}

\item{type}{Individual or Team work}

\item{pre}{Pre-amble for the assignment name, usually groupname + "-"}

\item{server}{The gitlab API server}
}
\description{
Create merge requests for each student/team
}
\details{
See \url{https://github.com/vnijs/gitgadget} for additional documentation
}

