% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{create_group}
\alias{create_group}
\title{Create a group on gitlab using the API}
\usage{
create_group(token, groupname = "", userfile = "", permission = 20,
  server = "https://gitlab.com/api/v4/")
}
\arguments{
\item{token}{Gitlab token}

\item{groupname}{Group to create on gitlab (defaults to user's namespace)}

\item{userfile}{A csv file with student information (i.e., username and token)}

\item{permission}{Permission setting for the group (default is 20, i.e., reporter)}

\item{server}{The gitlab API server}
}
\description{
Create a group on gitlab using the API
}
\details{
See \url{https://github.com/vnijs/gitgadget} for additional documentation
}
