% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{create_repo}
\alias{create_repo}
\title{Create the main repo from a local directory}
\usage{
create_repo(username, password, groupname, assignment, directory, pre = "",
  server = "https://gitlab.com/api/v3/")
}
\arguments{
\item{username}{Gitlab username}

\item{password}{Gitlab password}

\item{groupname}{Group to create on gitlab (defaults to user's namespace)}

\item{assignment}{Name of the assigment (repo)}

\item{directory}{Base directory for the repo. file.path(directory, assignment) should exist}

\item{pre}{Pre-amble for the assignment name, usually groupname + "-"}

\item{server}{The gitlab API server}
}
\description{
Create the main repo from a local directory
}
\details{
See \url{https://github.com/vnijs/gitgadget} for additional documentation
}
