% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{fetch_work}
\alias{fetch_work}
\title{Fetch all merge requests as local branches and link to a remote}
\usage{
fetch_work(token, assignment, server = "https://gitlab.com/api/v4/")
}
\arguments{
\item{token}{Gitlab token}

\item{assignment}{Name of the assigment (e.g., "class345/class345-assignment1")}

\item{server}{The gitlab API server}
}
\description{
Fetch all merge requests as local branches and link to a remote
}
\details{
See \url{https://github.com/vnijs/gitgadget} for additional documentation
}
