% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R, R/gadget.R
\docType{package}
\name{gitgadget}
\alias{gitgadget}
\title{gitgadget}
\usage{
gitgadget(port = get_port(), launch.browser = TRUE)
}
\arguments{
\item{port}{Port to use for the app}

\item{launch.browser}{Launch app in viewer (browsers) or only show the URL}
}
\details{
See \url{https://github.com/vnijs/gitgadget} for documentation
}
