% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R, R/gadget.R
\docType{package}
\name{gitgadget}
\alias{gitgadget}
\alias{gitgadget-package}
\alias{gitgadget}
\title{gitgadget}
\usage{
gitgadget()
}
\description{
gitgadget

Launch gitgadget in Rstudio viewer if available
}
\details{
See \url{https://github.com/vnijs/gitgadget} for documentation
}
